[mobile]refactor: refactor HomeFragment
