+++ mobile/app/src/main/java/com/thoughtworks/mobile/ui/fragments/home/HomeFragment.kt
@@ -9,7 +9,6 @@ import androidx.compose.ui.platform.ComposeView
 import androidx.compose.ui.platform.ViewCompositionStrategy
 import androidx.navigation.fragment.findNavController
 import com.thoughtworks.mobile.R
-import com.thoughtworks.mobile.ui.fragments.HomeViewModel
 import org.koin.androidx.viewmodel.ext.android.viewModel
 
 class HomeFragment : Fragment() {
